# specs/api/openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Kea Fabric API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/v1/admin/logs:
    get:
      operationId: get_admin_logs_api_v1_admin_logs_get
      parameters:
      - in: query
        name: service
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Service
      - in: query
        name: operation
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Operation
      - in: query
        name: subcategory
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Subcategory
      - in: query
        name: level
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Level
      - in: query
        name: mode
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mode
      - in: query
        name: from
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From
      - in: query
        name: to
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To
      - in: query
        name: cursor
        required: false
        schema:
          default: 0
          minimum: 0
          title: Cursor
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 500
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Admin Logs Api V1 Admin Logs Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Admin Logs
  /api/v1/admin/replication/summary:
    get:
      operationId: get_replication_summary_api_v1_admin_replication_summary_get
      parameters:
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Replication Summary Api V1 Admin Replication Summary
                  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Replication Summary
  /api/v1/dashboards/{dashboard_id}/layout:
    get:
      operationId: get_layout_api_v1_dashboards__dashboard_id__layout_get
      parameters:
      - in: path
        name: dashboard_id
        required: true
        schema:
          title: Dashboard Id
          type: string
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Layout Api V1 Dashboards  Dashboard Id  Layout
                  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Layout
    put:
      operationId: put_layout_api_v1_dashboards__dashboard_id__layout_put
      parameters:
      - in: path
        name: dashboard_id
        required: true
        schema:
          title: Dashboard Id
          type: string
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Put Layout
  /api/v1/dashboards/{dashboard_id}/layout/reset:
    post:
      description: Restore layout from read-only ``dashboard-layouts.orig.json`` on
        disk.
      operationId: reset_layout_api_v1_dashboards__dashboard_id__layout_reset_post
      parameters:
      - in: path
        name: dashboard_id
        required: true
        schema:
          title: Dashboard Id
          type: string
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Reset Layout Api V1 Dashboards  Dashboard Id  Layout
                  Reset Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reset Layout
  /api/v1/dashboards/{dashboard_id}/layout/save-file:
    post:
      description: Persist layout to the live store and write a timestamped JSON snapshot.
      operationId: save_dashboard_layout_file_api_v1_dashboards__dashboard_id__layout_save_file_post
      parameters:
      - in: path
        name: dashboard_id
        required: true
        schema:
          title: Dashboard Id
          type: string
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                title: Response Save Dashboard Layout File Api V1 Dashboards  Dashboard
                  Id  Layout Save File Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Save Dashboard Layout File
  /api/v1/dhcp/clients:
    get:
      operationId: get_clients_api_v1_dhcp_clients_get
      parameters:
      - in: query
        name: mock
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mock
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Clients Api V1 Dhcp Clients Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Clients
  /api/v1/dhcp/clients/{client_id}:
    patch:
      operationId: patch_client_api_v1_dhcp_clients__client_id__patch
      parameters:
      - in: path
        name: client_id
        required: true
        schema:
          title: Client Id
          type: string
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Patch Client Api V1 Dhcp Clients  Client Id  Patch
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch Client
  /api/v1/dhcp/pools:
    get:
      operationId: get_pools_api_v1_dhcp_pools_get
      parameters:
      - in: query
        name: mock
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mock
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Pools Api V1 Dhcp Pools Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Pools
  /api/v1/dhcp/reservations:
    get:
      operationId: get_reservations_api_v1_dhcp_reservations_get
      parameters:
      - in: query
        name: mock
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mock
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Reservations Api V1 Dhcp Reservations Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Reservations
  /api/v1/dhcp/reservations/{reservation_id}:
    patch:
      operationId: patch_reservation_api_v1_dhcp_reservations__reservation_id__patch
      parameters:
      - in: path
        name: reservation_id
        required: true
        schema:
          title: Reservation Id
          type: string
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Patch Reservation Api V1 Dhcp Reservations  Reservation
                  Id  Patch
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch Reservation
  /api/v1/discovery/records:
    get:
      operationId: get_discovery_records_api_v1_discovery_records_get
      parameters:
      - in: query
        name: mock
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mock
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Discovery Records Api V1 Discovery Records Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Discovery Records
  /api/v1/discovery/scan:
    get:
      operationId: get_discovery_scan_api_v1_discovery_scan_get
      parameters:
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Discovery Scan Api V1 Discovery Scan Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Discovery Scan
  /api/v1/discovery/scan/pause:
    post:
      operationId: post_discovery_pause_api_v1_discovery_scan_pause_post
      parameters:
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Post Discovery Pause Api V1 Discovery Scan Pause Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Post Discovery Pause
  /api/v1/events/stream:
    get:
      operationId: stream_events_api_v1_events_stream_get
      parameters:
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Stream Events
  /api/v1/health:
    get:
      operationId: get_health_api_v1_health_get
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Health Api V1 Health Get
                type: object
          description: Successful Response
      summary: Get Health
  /api/v1/meta:
    get:
      operationId: get_meta_api_v1_meta_get
      parameters:
      - in: query
        name: mock
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mock
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Meta Api V1 Meta Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Meta
  /api/v1/perf/summary:
    get:
      operationId: get_perf_api_v1_perf_summary_get
      parameters:
      - in: query
        name: mock
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mock
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Perf Api V1 Perf Summary Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Perf
  /api/v1/plugins:
    get:
      operationId: get_plugins_api_v1_plugins_get
      parameters:
      - in: query
        name: mock
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mock
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Plugins Api V1 Plugins Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Plugins